Q1/requirements.txt
# Core MCP Framework
fastmcp>=0.2.0

# Database & Data Models
sqlmodel>=0.0.8
sqlite3  # Built-in Python module

# Time Zone & Date Handling
pendulum>=3.0.0
pytz>=2023.3

# LLM Integration
openai>=1.14.0
anthropic>=0.25.0
google-generativeai>=0.5.0

# HTTP Client for API calls
httpx>=0.25.0

# Environment Variables
python-dotenv>=1.0.0

# Async Support
asyncio  # Built-in Python module

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-benchmark>=4.0.0

# Code Quality
ruff>=0.1.0

# Logging
structlog>=23.1.0

# UUID Generation
uuid  # Built-in Python module

# JSON Handling
json  # Built-in Python module

# Type Hints
typing-extensions>=4.7.0
